# Speed up example on plot_gradient_boosting_categorical.py 