#### Reference Issues/PRs
Contributes to #21598.

#### What does this implement/fix? Explain your changes.
In order to speed up the example, I did two main things : 

1. Decrease the number of columns of the chosen dataset from 80 to 20, keeping both numerical and categorical columns
2. Decrease the number of iterations of the `HistGradientBoostingRegressor` from 100 to 50

Please see the before-after for the plots, and the cProfile of the old and new versions: on my computer, the new version takes 4.5s vs 15.5s; the rank of the fit time and the error (plots) is the same for both figures: 

- cProfiles new version: `3894443 function calls (3808982 primitive calls) in 4.671 seconds`
- cProfiles old version: `6640438 function calls (6468083 primitive calls) in 15.580 seconds`

- plots new version : 
![Figure_1](https://user-images.githubusercontent.com/34744445/141309449-bf061340-bbc5-48bb-a004-43a4030f8566.png)
![Figure_2](https://user-images.githubusercontent.com/34744445/141309446-c8b45f42-addb-4696-87b2-2be85f8767d0.png)

- plots old version
![sphx_glr_plot_gradient_boosting_categorical_001](https://user-images.githubusercontent.com/34744445/141309543-0db2e5c9-7d23-42c1-bf09-fd4e1e4639dd.png)
![sphx_glr_plot_gradient_boosting_categorical_002](https://user-images.githubusercontent.com/34744445/141309549-46703640-fd36-4b41-8ad7-f5ca6a356988.png)