@TomDLT 2021-11-11
The CI error seems unrelated with this PR and caused by an unfortunate incompatibility between new numpy and old pandas (see https://github.com/numpy/numpy/issues/18355).