@pedugnat 2021-11-12
Thanks for your comments. 

I thought that reducing the number of columns would be the most promising option since 80 columns since unnecessary for the example to be relevant. I agree that the selection procedure adds too much complexity. 

Would you have any idea on either how to column selection in a clear and simple way (I was thinking we could manually select the columns and store them in a list at the beginning of the example but it doesn't look very clean either) or more generally how to reduce the training time?

Since the first part seems to be a bit underfitting with the new number of iterations, maybe I could increase the `learning_rate`  instead of increasing the number of iterations? This way the time gain would be kept.