@ogrisel 2021-11-12
Maybe you can subselect 10 categorical columns and 5 numerical columns manually by name and only change 1 line in the example? Try to focus on columns that are the most informative by running a permutation importance analysis on this dataset (outside of the example).