@ogrisel 2021-11-12
See: https://scikit-learn.org/stable/modules/permutation_importance.html