@pedugnat 2021-11-12
@ogrisel I ran the permutation importance you describe and selected 10 categorical variables and 10 numerical variables. 
I filtered the dataset to keep only these columns and here are the plots : 

They don't exactly match the previous ones, particularly because the version dropping categorical performs now significantly worse than the other versions (since there are fewer numerical variables, it may be harder for the 'dropped' version to learn enough). However, the benefit of using the native handling of categorical is shown. 

The new version now takes 4 seconds to run (instead of ~15).

What do you think?

![image](https://user-images.githubusercontent.com/34744445/141480240-ad9c1b63-08d8-4d82-9787-192fe78042ac.png)
![image](https://user-images.githubusercontent.com/34744445/141480256-a3d1fb99-8f22-4699-af20-9af9c7ab671b.png)