@ogrisel 2021-11-12
> They don't exactly match the previous ones, particularly because the version dropping categorical performs now significantly worse than the other versions (since there are fewer numerical variables, it may be harder for the 'dropped' version to learn enough).

It makes sense. I think this is not a problem.

> However, the benefit of using the native handling of categorical is shown.

And in my opinion, more importantly, the good predictive performance of OHE and Ordinal Encoder strategies when the model does not underfit it still there.

Can you confirm that you do not limit max_iter anymore in the first part of the example?

> The new version now takes 4 seconds to run (instead of ~15).
>
> What do you think?

This looks good, please feel free to update your PR accordingly.