@pedugnat 2021-11-12
Ok, I just updated the PR.

I can confirm that I restored `max_iter` to its default value.