@adrinjalali 2021-11-24
This is a nice improvement, but we have this warning:

> /home/circleci/project/sklearn/datasets/_openml.py:876: UserWarning: Version 1 of dataset ames-housing is inactive, meaning that issues have been found in the dataset. Try using a newer version from this URL: https://www.openml.org/data/v1/download/20649135/ames-housing.arff

It'd be nice if you could investigate the issue @pedugnat . Merging this one as is, since the warning is not caused by this PR.